Sprint Journal Formatting
Added some headings to sections such as CSC-324 Classmate Testing and Stakeholder Testing, I also moved the CIDER Lab to the bottom to have its own section to be easily navigated section-wise. Also still not entirely sure what to name Destany's section I would think it would be additional testing but then it should be moved between Stakeholder testing and CIDER LAB but it feels better as a chronological ordering for most of it.