Add Data Modeling Section to Requirements.md
I removed the original data modeling file as the new information on the sprint asked to have the data modeling in the Requirements.md. I added the written descriptions for our tables and database and the area to add the image. Still need to redraw the image for the schema.