Change dropExpiredEvents to read ids from JSON
Made changes to the dropExpiredEvents
now it reads in the json parses the structure, puts the ids into an array and then uses that array to fill our query. I also added a check on the array to ensure that we can just return if we have no events to delete to avoid a malformed query.