Implement MonthView for Calendar
Note:
This is primarily adapted from a tutorial I follow as Leah mentioned to not try to reinvent the wheel on this one. It does however use a lot of hard coded values that I will need to make some edits to, to make it more flexible eventually.

DONE:
- Visually outside of some hard coded values the MonthView page is done
- Logic for events and sorting by time will need to be implemented (Likely a later sprint)
- Update documentation

Commented out a selectedview on the homepage: not sure what it was doing there that should be removed or at least not tied to the same uses as the calendar.