Implement Verification UI for Sign up process
DONE:
I added a checker to see if the password matches our first password and that the first password has the requirements of a "good" password. I still need to implement the logic for disabling the button until all the fields are filled. Also made them into outlined text fields because they look better.