Refactor Org Card and Settings Page to follow theme
In short, I made it so that all components on the org card and the settings page followed our theme. In hopes that this makes it easier for us to change it quickly when we settle on a color scheme. Additionally, made changes to the settings page to follow the suggestions Michael gave on Discord.