Implement a local database
Done:
Implemented the general beginning of a local database. Currently does not persist favorite events for some reason if the app is closed likely because it refreshes from the api on build and thus changes them back, since we have no way of saving this remotely yet. Overall, this was a headache, but I'm happy to have something to show for it finally.