Fix Merge Conflict issues and Unfavoriting UI bug
I fixed the merge conflicts that we had so now all items work as intended. I also changed where the cards derive their border and background color so that they are the correct colors for light and dark mode while still keeping that background color and border. Additionally changed the tags to fit the screen of any screen for android as we intended visually. Finally fixed the bug with the favorites so now unfavoriting no longer has the issue of a stale visual state.