Fix Verification Navigation bug
This bug occurred because the way the android app runs is that while we don’t kill the app when the user leaves to check the email the app still has to decide what page to navigate the user to when the come back to it and before we only had the option of if they were logged in that they went to the home page and if not they went to the welcome screen. I introduced another route that maintains important verification-related fields, so the app can now check this condition and ensure it takes the user to the correct process. Thus, we have fixed this issue.

TLDR: By introducing a new possible first screen display on load route we are able to show the verification page even when a user leaves the app to check their email and comes back.