Check for existing account when changing email
We do checks both on the initial request, and on the verification step.
I think there is a possibility of a race condition still because the
database function `changeUserEmail` is not atomic with respect to
requests, but there is still no way for an illegal state to happen--it
may just cause an unknown error client side.