Add test for changing email to existing account
The race condition I identified seems to be impossible to autotest,
because it relies entirely on something that happens with email
verification.

Normally the solution would be to add it to the manual testing plan, but
we don't have a way to do this--accessing two emails is already so much
of a stretch. I think we may need to just let this beast lie.